refactor(core): move fetch configuration to manager-config
