+++ packages/manager/modules/config/package.json
@@ -27,7 +27,8 @@
     "start:watch": "lerna exec --stream --parallel --scope='@ovh-ux/manager-config' --include-dependencies -- yarn run dev:watch"
   },
   "dependencies": {
-    "lodash-es": "^4.17.15"
+    "lodash-es": "^4.17.15",
+    "whatwg-fetch": "^3.0.0"
   },
   "devDependencies": {
     "@ovh-ux/component-rollup-config": "^7.0.0"
