+++ packages/manager/modules/config/src/index.js
@@ -1,4 +1,5 @@
 import _Environment from './environment';
+import 'whatwg-fetch';
 
 import {
   convertLanguageFromOVHToBCP47 as _convertLanguageFromOVHToBCP47,
@@ -11,6 +12,12 @@ import {
   localeStorageKey as _localeStorageKey,
 } from './locale/locale.constants';
 
+export const HOSTNAME_REGIONS = {
+  'www.ovh.com': 'EU',
+  'ca.ovh.com': 'CA',
+  'us.ovhcloud.com': 'US',
+};
+
 export const Environment = _Environment;
 
 export const convertLanguageFromOVHToBCP47 = _convertLanguageFromOVHToBCP47;
@@ -19,10 +26,38 @@ export const findLanguage = _findLanguage;
 export const LANGUAGES = _LANGUAGES;
 export const localeStorageKey = _localeStorageKey;
 
+export const fetchConfiguration = () => {
+  return fetch('/engine/2api/configuration', {
+    headers: {
+      'Content-Type': 'application/json;charset=utf-8',
+      Accept: 'application/json',
+    },
+    credentials: 'same-origin',
+  })
+    .then((response) => {
+      if (response.status === 401) {
+        window.location.assign(
+          `/auth?action=disconnect&onsuccess=${encodeURIComponent(
+            window.location.href,
+          )}`,
+        );
+      }
+      return response.json();
+    })
+    .then((config) => {
+      _Environment.setRegion(config.region);
+      return config;
+    })
+    .catch(() => ({
+      region: HOSTNAME_REGIONS[window.location.hostname],
+    }));
+};
+
 export default {
   convertLanguageFromOVHToBCP47,
   Environment,
   detectUserLocale,
+  fetchConfiguration,
   findLanguage,
   LANGUAGES,
   localeStorageKey, // TODO(@antleblanc): Remove it on February 4th, 2021.
